Split out response_objects into their own namespace and files.
