+++ api/namex/resources/auto_analyse/response_objects/name_analysis_response.py
@@ -1,0 +1,7 @@
+from . import Serializable
+
+
+class NameAnalysisResponse(Serializable):
+    def __init__(self, **kwargs):
+        self.status = kwargs['status'] or ''
+        self.issues = kwargs['issues'] or []
